requirements.txt
numpy
scipy
pandas
jax
jaxlib
pyscf
tensorcircuit-nightly
openfermion
qiskit
pylatexenc
SciencePlots
noisyopt
renormalizer==0.0.8
